+++ c/src/lib.rs
@@ -1,4 +1,5 @@
 mod cfile;
+pub mod compat;
 pub mod decode;
 pub mod metadata;
 pub mod text_serialization;